Move the downgrade-related stuff to its own folder.
This is done in preparation of turning the downgrade functionallity
"off", as we won't be using it anymore.

This commit also removes the `usingDowngradePath` value from the state,
as it wasn't even used and to make the codebase a bit less polluted with
stuff that probably won't be used for a while. If it ever turns out in
future that we need the downgrade functionality again, the state can be
put back in, if needed.
