+++ client/downgrade/patch-starcraft.js
@@ -4,7 +4,7 @@ import glob from 'glob'
 import childProcess from 'child_process'
 import mkdirp from 'mkdirp'
 import thenify from 'thenify'
-import { EXE_HASHES_1161, STORM_HASHES_1161 } from './check-starcraft-path'
+import { EXE_HASHES_1161, STORM_HASHES_1161 } from '../starcraft/check-starcraft-path'
 import { streamEndPromise, streamFinishPromise } from '../../app/common/async/stream-promise'
 import getFileHash from '../../app/common/get-file-hash'
 import checkFileExists from '../../app/common/check-file-exists'
