+++ client/reducers.js
@@ -4,6 +4,7 @@ export { default as auth } from './auth/auth-reducer'
 export { default as bans } from './admin/bans-reducer'
 export { default as chat } from './chat/chat-reducer'
 export { default as dialog } from './dialogs/dialog-reducer'
+export { default as downgrade } from './downgrade/downgrade-reducer'
 export { default as fileBrowser } from './file-browser/file-browser-reducer'
 export { default as gameClient } from './lobbies/game-client-reducer'
 export { default as gameplayActivity } from './gameplay-activity/gameplay-activity-reducer'
